update library loading
